Add Mondrian user-defined func for rates
This adds a population calculator for rate calculation, as well as table
definitions, etc. to make it work. It also adds a sample use of the function in
the OLAP schema, as well as a sample of a PostgreSQL aggregate as a Mondrian
aggregator. It removes the old UDF code, and uses a NetBeans-friendly structure
for the new code.