supports :x defaults evaulated at load time
In the initial revision of supports :something (d2ecd06)
the defaults which were class evaluated did not get evaluated until the
first time class was instanitated.  This meant if you were to call
Event.new.supports_something? before calling EventSubclass.new which had
support :something defined, there would be a method missing error.

As a work around, we now define what functionality is supported so we
can define methods on the super class when the class is loaded.